+++ packages/expo-file-system/src/FileSystem.ts
@@ -1,3 +1,5 @@
+import { Blob as ExpoBlob } from 'expo-blob';
+
 import ExpoFileSystem from './ExpoFileSystem';
 import type { DownloadOptions, PathInfo } from './ExpoFileSystem.types';
 import { PathUtilities } from './pathUtilities';
@@ -68,7 +70,8 @@ export class Paths extends PathUtilities {
  * const file = new File(Paths.cache, "subdirName", "file.txt");
  * ```
  */
-export class File extends ExpoFileSystem.FileSystemFile implements Blob {
+
+export class File extends ExpoFileSystem.FileSystemFile {
   /**
    * Creates an instance of a file. It can be created for any path, and does not need to exist on the filesystem during creation.
    *
@@ -123,8 +126,12 @@ export class File extends ExpoFileSystem.FileSystemFile implements Blob {
     return this.readableStream();
   }
 
-  slice(start?: number, end?: number, contentType?: string): Blob {
-    return new Blob([this.bytesSync().slice(start, end)], { type: contentType });
+  slice(start?: number, end?: number, contentType?: string): ExpoBlob {
+    return new ExpoBlob([new Uint8Array(this.bytesSync())], { type: contentType }).slice(
+      start,
+      end,
+      contentType
+    );
   }
 }